requirements.txt
playwright==1.30.0
pytest
python-dotenv
